populate container commands for scan actions
this is a bit of a hack to patch a bug in commandUiLauncher.js, where
the xsiType parameter for the available commands endpoint is contructed
like this:

var xsiScanType = xsiType.replace('Session','Scan');

in our case this will not work, since our session data type is called
"sleepResearchSessionData" and our scan data type is called "psgScanData".

the above line assumes our scan data type is called
"sleepResearchScanData".